+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Add support for [Iceberg connector](https://trino.io/docs/current/connector/iceberg.html) ([#285]).
+
+[#285]: https://github.com/stackabletech/trino-operator/pull/285
+
 ## [0.6.0] - 2022-09-08
 
 ### Changed